Avoid Namespace in Imports field not imported from
Relates to #108

Avoid note as in https://github.com/hadley/r-pkgs/issues/203
While inelegant, this seems to be the recommended solution.
It solves the note in the subject of this commit for most
of the problematic platforms. There is still one platform
I'm not sure about (details below). But in any case, this
commit brings us closer to the goal.

DETAILS

CRAN's check results at
https://cran-archive.r-project.org/web/checks/2020/2020-02-20_check_results_datos.html

I no longer see the note on rhub with:
Platform:	Fedora Linux, R-devel, clang, gfortran
https://builder.r-hub.io/status/datos_0.2.0.tar.gz-d06dd13075724881be4ee9da953cd7ce

But seems to persist on rhub with
Platform:	Debian Linux, R-devel, GCC
https://builder.r-hub.io/status/datos_0.2.0.tar.gz-109732efb4d34774b3d1124954b63ba4
Maybe this result is from an older run. The results take a long
time to arrive and older and later runs might get mixed.